refactor: Extract app layout widget core runtime api
