+++ src/app-layout/visual-refresh-toolbar/state/use-app-layout.tsx
@@ -14,7 +14,6 @@ import globalVars from '../../../internal/styles/global-vars';
 import { getSplitPanelDefaultSize } from '../../../split-panel/utils/size-utils';
 import { AppLayoutProps } from '../../interfaces';
 import { SplitPanelProviderProps } from '../../split-panel';
-import { useAiDrawer } from '../../utils/use-ai-drawer';
 import { MIN_DRAWER_SIZE, OnChangeParams, useDrawers } from '../../utils/use-drawers';
 import { useAsyncFocusControl, useMultipleFocusControl } from '../../utils/use-focus-control';
 import { useGlobalScrollPadding } from '../../utils/use-global-scroll-padding';
@@ -27,6 +26,8 @@ import {
 } from '../compute-layout';
 import { AppLayoutState } from '../interfaces';
 import { AppLayoutInternalProps, AppLayoutInternals } from '../interfaces';
+import { useAiDrawer } from './use-ai-drawer';
+import { useWidgetMessages } from './use-widget-messages';
 
 export const useAppLayout = (
   hasToolbar: boolean,
@@ -145,6 +146,7 @@ export const useAppLayout = (
   });
   const {
     aiDrawer,
+    aiDrawerMessageHandler,
     onActiveAiDrawerChange,
     activeAiDrawer,
     activeAiDrawerId,
@@ -152,11 +154,11 @@ export const useAppLayout = (
     minAiDrawerSize,
     onActiveAiDrawerResize,
   } = useAiDrawer({
-    isEnabled: hasToolbar,
     onAiDrawerFocus: () => aiDrawerFocusControl.setFocus(),
     expandedDrawerId,
     setExpandedDrawerId,
   });
+  useWidgetMessages(hasToolbar, message => aiDrawerMessageHandler(message));
   const aiDrawerFocusControl = useAsyncFocusControl(!!activeAiDrawer?.id, true, activeAiDrawer?.id);
 
   const onActiveDrawerChangeHandler = (
