+++ src/internal/plugins/widget.ts
@@ -1,4 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 export * from './widget/interfaces';
-export { registerLeftDrawer, updateDrawer, isAppLayoutReady, whenAppLayoutReady } from './widget/internal';
+export { isAppLayoutReady, whenAppLayoutReady } from './widget/core';
+export { registerLeftDrawer, updateDrawer } from './widget/index';
